+++ internal/app.go
@@ -16,6 +16,7 @@ import (
 	"github.com/goverland-labs/core-storage/internal/dao"
 	"github.com/goverland-labs/core-storage/internal/events"
 	"github.com/goverland-labs/core-storage/internal/proposal"
+	"github.com/goverland-labs/core-storage/internal/vote"
 	"github.com/goverland-labs/core-storage/pkg/health"
 	"github.com/goverland-labs/core-storage/pkg/prometheus"
 )
@@ -115,6 +116,11 @@ func (a *Application) initServices() error {
 		return fmt.Errorf("init proposal: %w", err)
 	}
 
+	err = a.initVote(nc)
+	if err != nil {
+		return fmt.Errorf("init vote: %w", err)
+	}
+
 	return nil
 }
 
@@ -150,13 +156,30 @@ func (a *Application) initProposal(nc *nats.Conn, pb *communicate.Publisher) err
 
 	cs, err := proposal.NewConsumer(nc, service)
 	if err != nil {
-		return fmt.Errorf("dao consumer: %w", err)
+		return fmt.Errorf("proposal consumer: %w", err)
 	}
 
 	a.manager.AddWorker(process.NewCallbackWorker("proposal-consumer", cs.Start))
 
 	vw := proposal.NewVotingWorker(service)
 	a.manager.AddWorker(process.NewCallbackWorker("voting-worker", vw.Start))
+
+	return nil
+}
+
+func (a *Application) initVote(nc *nats.Conn) error {
+	repo := vote.NewRepo(a.db)
+	service, err := vote.NewService(repo)
+	if err != nil {
+		return fmt.Errorf("vote service: %w", err)
+	}
+
+	cs, err := vote.NewConsumer(nc, service)
+	if err != nil {
+		return fmt.Errorf("vote consumer: %w", err)
+	}
+
+	a.manager.AddWorker(process.NewCallbackWorker("vote-consumer", cs.Start))
 
 	return nil
 }